Moved suntime to /utils/
